### 2021-05-21 #3302 @whedon
[PRE REVIEW]: Nominally: A Name Parser for Record Linkage
**Submitting author:** @vaneseltine (<a href="http://orcid.org/0000-0002-9520-1360">Matthew VanEseltine</a>)
**Repository:** <a href="https://github.com/vaneseltine/nominally" target ="_blank">https://github.com/vaneseltine/nominally</a>
**Version:** v1.1.0
