### 2021-05-21 #3302 @kyleniemeyer
Hi @vaneseltine, thanks for your interest in JOSS. Due to low editor availability in this area right now, we have to put your submission on a waitlist. Once an appropriate editor becomes available, we will assign them and move to finding reviewers. In the meantime, if you could suggest some potential reviewers, that would help speed up the process. Thanks!
