### 2021-05-28 257e730
Improve noxfile classifier extraction
