### 2021-06-18 #3302 @danielskatz
@majensen - thanks!  When you take something from the waiting list, can you also remove the tag, just to keep things clean?
