### 2021-07-01 #3302 @majensen
@sap218 that is great. Please have a look at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and https://joss.readthedocs.io/en/latest/review_criteria.html. A checklist that covers the review areas will show up in the review issue (which will be created once I put the paper in review.) You can check these off as you go through at the top of the review thread.
Very much appreciated!
