### 2021-07-05 vaneseltine/nominally#54 @sap218
parse_name() issues with title "miss"
Currently reviewing as part of https://github.com/openjournals/joss-reviews/issues/3440

The examples provided work as expected - however when using my own examples:

`parsed = nominally.parse_name("Miss Samantha Clare Pendleton")` - output not as expected and title blank / instead `"first":"miss"`

`parsed = nominally.parse_name("Ms Samantha Clare Pendleton")` - output as expected

`parsed = nominally.parse_name("Mrs Samantha Clare Pendleton")` - output as expected

All outputs are as expected except for "Miss" which parse_name() function cannot identify as a title
