### 2021-07-05 #3440 @sap218
1. Regarding "Contribution and authorship" - the submitting author (@vaneseltine) has made major contributions to the software and is an author on the paper. However I did notice in the whedon software report (https://github.com/openjournals/joss-reviews/issues/3440#issuecomment-872466076) that Derek Gulbranson made ~25% of changes for Nominally - I did quickly check the commit history and noticed Derek last committed around June 2019. Just wondering if authorship should be updated? If not, I'll be happy to check that off and continue :)

2. Installation / documentation - I like the set-up of the GitHub Repo for Nominally - however there lacks installation instructions. I did find them on ReadTheDocs (https://nominally.readthedocs.io/en/stable/use.html#installation) so I think either including installation instructions in the repo README or perhaps in the repo there should be a link to the ReadTheDocs manual! Additionally, not a major issue but the examples on the repo and the ReadTheDocs are different, suggestion would be to make them the same/consistent.

3. First error I came across was Nominally parse_name() function cannot identify "miss" as a title. As soon as this is addressed (https://github.com/vaneseltine/nominally/issues/54) I will continue with the review! :)
