### 2021-07-11 vaneseltine/nominally#54 @vaneseltine
Thanks for filing this issue! In reviewing the title list, I also added [Mx](https://en.wikipedia.org/wiki/Mx_(title)) to bring the total to `TITLES = {"dr", "miss", "mr", "mrs", "ms", "mx"}`

Additional thoughts: There are many additional titles that could be added (see Wikipedia on [English honorifics](https://en.wikipedia.org/wiki/English_honorifics) or [Titles in English-speaking areas](https://en.wikipedia.org/wiki/Title#Titles_in_English-speaking_areas); cf. [python-nameparser's extensive list](https://github.com/derek73/python-nameparser/blob/master/nameparser/config/titles.py)) but I'd like to keep the false-positive title labeling low. In record linkage across different data sources, titles are unlikely to have a strong signal -- it would be a rare circumstance where Mr or Dr or Miss are helpful determinants of matching. This short list can be safely excluded entirely. Better to keep a title as an extraneous name than to throw away a potentially useful name such as [Sir](https://www.wolframalpha.com/input/?i=name+sir).
