### 2021-07-12 #3440 @sap218
@vaneseltine @majensen

Since my issue was closed I have continued with the review - I have ticked off more tasks :)

Documentation
1. A statement of need is needed in the docs (even if very brief)
2. Continuing on from my last comment (2) I think a simple line of installation instructions in the main repo would be great :)
3. There are numerous source of documentation - there is the Nominally main repo, ReadTheDocs, and a nominally-examples repo - each have examples not the same as others (I am ok with this - but would be nice if examples were consistent across all). I would perhaps add more examples to the examples repo README

To conclude overall: a statement of need, installation instructions, examples of use - are all needed in each types of docs you provide - I can finish off with the software paper as soon as these are addressed :)
