### 2021-07-16 #3440 @vaneseltine
> 1. Regarding "Contribution and authorship" - the submitting author (@vaneseltine) has made major contributions to the software and is an author on the paper. However I did notice in the whedon software report ([[REVIEW]: Nominally: A Name Parser for Record Linkage #3440 (comment)](https://github.com/openjournals/joss-reviews/issues/3440#issuecomment-872466076)) that Derek Gulbranson made ~25% of changes for Nominally - I did quickly check the commit history and noticed Derek last committed around June 2019. Just wondering if authorship should be updated? If not, I'll be happy to check that off and continue :)

Thanks for raising this, as it's important to me that this is all clear and correct. Nominally started as a fork of Derek Gulbranson's python-nameparser project (https://github.com/derek73/python-nameparser). As I started digging into the package more, I developed nominally into an entirely separate package with different core processes and internals, working in the same problem space. As the whedeon report indicates, very few lines (it counted 10) actually remain from his work; my own git blame yielded:

```
.gitignore

bf0c62c7 (Derek Gulbranson 2014-05-06 22:55:47 -0700 36) *.egg
b4d232c7 (Derek Gulbranson 2014-05-16 18:30:35 -0700 52)
280895bd (Derek Gulbranson 2011-02-03 18:22:15 -0800 1) include LICENSE

nameparser/parser.py

5f3da6c8 (Derek Gulbranson 2016-08-04 21:13:11 -0700   5)
efe25583 (Derek Gulbranson 2018-08-31 17:18:15 -0700  57)
7446b6ab (Derek Gulbranson 2014-05-15 04:33:10 -0700 125)
5a78cf63 (Derek Gulbranson 2014-04-01 19:16:53 -0700 173)         """
03e580c0 (Derek Gulbranson 2018-08-30 12:49:02 -0700 175)         nicknames list. This happens before any other processing of the name.
03e580c0 (Derek Gulbranson 2018-08-30 12:49:02 -0700 180)         """
```
That does undersell the legacy, however! I'd converted much of the python-nameparser test database from its single-file _nose_-based suite https://github.com/derek73/python-nameparser/blob/master/tests.py into a parameterized _pytest_-based system. (That was actually the first step I took in working on this, see https://github.com/derek73/python-nameparser/pull/91.)

I include python-nameparser in an acknowledgements section of both this manuscript and the repository. To me, this seems an appropriate way to strike the right balance; I don't want to erase the history of the project (literally and figuratively), but nominally as it stands today is my own work. I'm happy to discuss further, and if JOSS has standards regarding authorship with forks, this would be helpful for me to review.
