### 2021-07-19 #3440 @sap218
@vaneseltine Thank you for the reply!

I have gone through the checklist and I'll put my final reviewer comments below :)

**General Checks**
All good!

**Functionality**
All good!

**Documentation**
1. A statement of need should included the the main git repo (just a shortened version from the paper)
2. Installation instructions are needed in the main git repo
3. Include instructions how to run tests in the main git repo
4. Example usage: you include examples in the main repo / ReadTheDocs / software paper and even though I checked this off the time - it would be nice to have the examples the same (but don't worry too much about this one)

**Software Paper**
All good! I have one minor comment though:
1. Line 30 when you mentioned "a rule-based system" - maybe a line afterwards about what this rule-based system is (but I wouldn't worry about this comment since you do cite it!)
