### 2021-08-23 #3440 @sara-02
@majensen @@vaneseltine I have completed my review (3 points remaining, discussed below)
1. For testing, while the test folder is visible in the repo, I do not see any instructions for setting up and running the tests. Adding commands in the readme will be helpful here.
2. For Community guidelines, a section on the community contributor covering the 3 aspects required by the JOSS guideline should be added in the readme.
3.  A direct link to the documentation page either in the readme or in the repo URL should be added, I landed on the docs via the PyPI page of the package.
4.  The comparision with other softwares, is clubed with statement of need section in the paper, it can be a separate section (not a strict requirement but makes the paper more structred).
