### 2021-10-01 8c7dcb3
Paper - split comparison into new section
