### 2021-10-01 9dbebde
Improvements from JOSS review
