### 2021-10-01 #3440 @vaneseltine
Thanks to both reviewers for these great suggestions!

The documents in the core repository have now been improved based on several (combined) suggestions in https://github.com/vaneseltine/nominally/commit/9dbebde85c83606141f0189a3ae69332111cfd31:

1. A statement of need. I incorporated language from the paper on the difficulties of name records into README intro, in order to motivate the usefulness of the software there in the introduction.

1. A link to the full documentation. I added a README section, "Documentation."

1. Installation instructions. I added a README section, "Installation."

1. Following the JOSS community guidelines to show how to (a) contribute to the software, (b) report issues or problems with the software, and (c) seek support. I added a README section, "Community."

1. Instructions to run tests. I improved the description in CONTRIBUTING.md and refer users there for tests in the new "Community" section of the README.

Additionally, @sap218 noted the divergence in various examples used in the Github repository, the JOSS manuscript, and the Read The Docs documentation. I did update the Github README to use the same core example ("Vimes, jr, Mr. Samuel 'Sam'") as the manuscript. The Read the Docs documentation intentionally uses a few different names across various examples, so I did leave it as-is, if that seems reasonable. The README was updated to use this example in https://github.com/vaneseltine/nominally/commit/869e8c621bf2725eb6528542c5bcd36742685ef2

And finally, following @sara-02's suggestion to separate the comparison with other software in the paper, I have added a section just so in the manuscript. https://github.com/vaneseltine/nominally/commit/8c7dcb305356ddfb29e6477ca15ccc4519f4d087

I think that together, these updates address the reviewers' concerns. Thanks again!
