### 2021-10-11 #3440 @majensen
@vaneseltine This is looking very good. I would make two editorial changes:

- Line 26 of proof: ‘okina is actually spelled with an ‘okina (https://en.wikipedia.org/wiki/%CA%BBOkina); also in README.md
- Line 81: I would credit [DataMade](https://datamade.us/) rather than the AJC for probablepeople, since it looks like they did the coding. Up to you.

Just FYI, not for the paper:
A basic "nox" leads to flake8 reporting a bunch of style issues and then a bork.
If someone wants to just see if the code tests work, I would advise devs to do
```
virtualenv .venv
source .venv/bin/activate
pip3 -r requirements/test.txt
pytest
```
(which succeeds, I am pleased to note).

Once you have made any last changes, can you create an open archive of the repository in Zenodo or similar (here's a [tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729)) and then report the DOI you receive in this thread? Note that JOSS prefers the archive to be named with the name of the paper and the author list to be the same as that of the paper. Much appreciated!
