### 2021-10-11 vaneseltine/nominally#55 @vaneseltine
nox fails flake8 on fresh install
https://github.com/openjournals/joss-reviews/issues/3440#issuecomment-939600112

> A basic "nox" leads to flake8 reporting a bunch of style issues and then a bork.
> If someone wants to just see if the code tests work, I would advise devs to do
>
> virtualenv .venv
> source .venv/bin/activate
> pip3 -r requirements/test.txt
> pytest
>
> (which succeeds, I am pleased to note).
