+++ paper.bib
@@ -39,7 +39,7 @@
 }
 
 @misc{ probablepeople,
-  author = {{The Atlanta Journal Constitution}},
+  author = {{DataMade}},
   title = {probablepeople},
   year = {2019},
   publisher = {GitHub},
