### 2021-10-11 b6b91d3
Default nox to 'test or lint'
