### 2021-10-11 #3440 @vaneseltine
Hrm, looks like the the ʻokina (U+02BB) does not end up properly rendered in the PDF.
![image](https://user-images.githubusercontent.com/20784848/136826990-bd01e00d-094e-48b2-98d4-6ceb075e15d9.png)
