### 2021-10-11 #3440 @majensen
Shoot. Well, I guess we can fake it with U+2018 (left single quotation mark).
