### 2021-10-11 af19439
Paper - sub U+2018 for U+02BB okina
